fixes after code review
