+++ frontend/src/components/client/CompleteAssignments/CompleteAssignments.jsx
@@ -21,7 +21,7 @@ import decodeStyledText from '../../../service/decodeStyledText';
 import Modal from '../../modals/Modal/Modal';
 import AssignmentNotComplete from '../../modals/Notifications/assignmentNotComplete';
 import AssignmentExit from '../../modals/Notifications/assgnmentExit';
-import { minMobWidth, maxMobWidth } from '../../../utils/constants';
+import useMobileWidth from '../../../utils/hook/useMobileWidth';
 
 function CompleteAssignments() {
   const location = useLocation();
@@ -311,24 +311,7 @@ function CompleteAssignments() {
     }
   }
 
-  const [isMobileWidth, setIsMobile] = useState(false);
-
-  useEffect(() => {
-    const handleResize = () => {
-      const width = window.innerWidth;
-      if (width >= minMobWidth && width <= maxMobWidth) {
-        setIsMobile(true);
-      } else {
-        setIsMobile(false);
-      }
-    };
-
-    // Sets the initial state based on the current window size
-    handleResize();
-
-    window.addEventListener('resize', handleResize);
-    return () => window.removeEventListener('resize', handleResize);
-  }, []);
+  const isMobileWidth = useMobileWidth();
 
   return isRateTask ? (
     <>
