+++ frontend/src/components/client/DiaryPage/DiaryPageClient.jsx
@@ -32,25 +32,6 @@ export default function DiaryPageContentClient({ diary, type }) {
     mode: 'all',
   });
 
-  const [isMobileWidth, setIsMobile] = useState(false);
-
-  useEffect(() => {
-    const handleResize = () => {
-      const width = window.innerWidth;
-      if (width >= minMobWidth && width <= maxMobWidth) {
-        setIsMobile(true);
-      } else {
-        setIsMobile(false);
-      }
-    };
-
-    // Sets the initial state based on the current window size
-    handleResize();
-
-    window.addEventListener('resize', handleResize);
-    return () => window.removeEventListener('resize', handleResize);
-  }, []);
-
   const onSubmit = async (data) => {
     console.log(data);
